// lib/enums/BrdBlm_eTextureType.ts
/** ---------------------------------------------------------------------------
 * @module [BrdBlm]
 * @author [APG] Angeli Paolo Giusto
 * @version 0.1 APG 20240214
 * @version 0.2 APG 20240323 Spostato nel suo file dedicato
 * ----------------------------------------------------------------------------
 */


/**
 * Tipi di texture usati nelle {@link BrdBlm_ITextureDef} dei {@link BrdBlm_IMaterialDef}
 */
export enum BrdBlm_eTextureType {
    ALBEDO = '_AbedoMap',
    BUMP = '_BumpMap',
    NORMAL = '_NormalMap',
    ALPHA = '_AlphaMap',
}



/*! ---------------------------------------------------------------------------
 * @copyright Breda Sistemi industriali S.p.A., 2023 - http://bredasys.com
 * All rights reserved 
 * @licence You cannot host, display, distribute or share this Work in any 
 * form, both physical and digital. You cannot use this Work in any commercial
 * or non-commercial product, website or project. You cannot sell this Work
 * and you cannot mint an NFTs out of it.
 * --------------------------------------------------------------------------- 
 */